Implement the random-backend-selection proxy functionality
Create a httputil.ReverseProxy director that randomly selects from a
backend at runtime. Maintain a map from service domains to the
reverse proxy for each domain and its backends.

When handling a request check the "s" parameter for the service's
domain, and proxy the request using the reverse proxy for that domain.

Update the README accordingly.