Add tests for health checking functionality
Add a health checker function to the Proxy struct, and use that instead
of the hard-coded health check. Provide a default health checker function
that always returns ok.

Use this to test the health checking feature with two tests, one that
verifies that the default works, and one that verifies that an "always
fail" health checker is also called correctly.